fix up how we invoke benchmarks
