+++ src/tests.rs
@@ -1,7 +1,5 @@
 use cm;
 use html;
-#[cfg(feature = "benchmarks")]
-use test::Bencher;
 use timebomb::timeout_ms;
 use {parse_document, Arena, ComrakOptions};
 
@@ -60,23 +58,6 @@ macro_rules! html_opts {
             $(opts.$optclass.$optname = true;)*
         });
     };
-}
-
-#[cfg(feature = "benchmarks")]
-#[cfg_attr(feature = "benchmarks", bench)]
-fn bench_progit(b: &mut Bencher) {
-    use std::fs::File;
-    use std::io::Read;
-
-    let mut file = File::open("script/progit.md").unwrap();
-    let mut s = String::with_capacity(524288);
-    file.read_to_string(&mut s).unwrap();
-    b.iter(|| {
-        let arena = Arena::new();
-        let root = parse_document(&arena, &s, &ComrakOptions::default());
-        let mut output = vec![];
-        html::format_document(root, &ComrakOptions::default(), &mut output).unwrap()
-    });
 }
 
 #[test]
